- Added ImageServlet to so the uploaded images will not be lost on redeploy - Updated the fileuploadbean and fileupload.jsp to use new method - updated search and map popup images to use new method
To convert to new method, add the "ImageServlet" to your url, for example:

http://localhost:8080/petstore/ImageServlet/images/bird1.gif

The uploaded images uses the same format, for example:

http://localhost:8080/petstore/ImageServlet/images/newly_uploaded_image.gif


The old url would have been:

http://localhost:8080/petstore/images/bird1.gif


The ImageServlet first checks to default location and then checks the upload location for the image.

If the image isn't found a severe log message is printed to the log.




Issue number:
Obtained from:
Submitted by:
Reviewed by: